requirements.txt
anyio>=3.3.4
fastapi>=0.70.0 
idna>=3.3 
pydantic>=1.8.2 
sniffio>=1.2.0
starlette>=0.16.0
typing-extensions>=4.0.0
unicorn>=1.0.3
psycopg2-binary-2.9.2

